Config+Auth: Add flags to log unauthorized requests
This patch adds new command line flags in order to support logging of
unauthorized requests to the server. The flag `--log-auth-failure` enables
the logging and uses the remote address of the request as the default for
the logged ip. If the server is used behind a reverse proxy for, `--ip-header`
can be used to specify a header like "X-Forwarded-For" to be used for logging
the ip.